+++ backend/src/main/kotlin/com/nom/api/domain/menu/ports/in/UpdateMenuItemUseCase.kt
@@ -1,7 +1,6 @@
 package com.nom.api.domain.menu.ports.`in`
 
 import com.nom.api.domain.menu.entities.MenuItem
-import java.math.BigDecimal
 
 interface UpdateMenuItemUseCase {
     suspend fun execute(request: UpdateMenuItemRequest): MenuItem
@@ -12,7 +11,7 @@ data class UpdateMenuItemRequest(
     val menuItemId: String,
     val name: String,
     val description: String? = null,
-    val price: BigDecimal,
+    val price: Long,
     val allergens: List<String> = emptyList(),
     val imageUrl: String? = null
 )
