+++ backend/src/main/kotlin/com/nom/api/domain/order/entities/OrderItem.kt
@@ -1,10 +1,9 @@
 package com.nom.api.domain.order.entities
 
-import java.math.BigDecimal
 
 data class OrderItem(
     val menuItemId: String,
     val name: String,
-    val price: BigDecimal,
+    val price: Long,
     val quantity: Int
 )
